bedrockrs Full Dependency
Tried to import a specific implementation for the DB usage, but it doesn't quite make sense to me. Going to see if the API is more straightforward with the full library itself. I don't want to include extra dependencies I don't use though. Rust doesn't have tree-shaking does it? I would think it does, but maybe not. It could just be that it always builds everything in dev, but only uses it in prod?