Logging DB Values
I started debugging the contents that are read with the iterator, it does appear to be reading them correctly, which is great! I am curious about whether Bedrock worlds still use zlib/snappy for the DB backing under the hood, or if they aren't using it anymore. I thought I just read not that long ago that they don't use it anymore, but I feel like that couldn't be true, why wouldn't you want to use compression? It does seem to be working with the same values though if I don't enable the snappy compression flag. So, I'll leave it on for now though, then once things are more established with parsing the data I am just now getting access to, then I can more properly validate whether it's using compression itself.

Gotta say, I love that you can define tuple destructured variables, and separately define their mutability, AND describe their types with an annotation, both of which are awesome. I wish in JS/TS you could define one as const one as let, as well as set an annotation on the destructuring.

https://doc.rust-lang.org/std/vec/struct.Vec.html
https://stackoverflow.com/questions/39785597/how-do-i-get-a-slice-of-a-vect-in-rust
https://doc.rust-lang.org/std/vec/struct.Vec.html#method.truncate
https://stackoverflow.com/questions/66073157/why-can-i-start-a-slice-past-the-end-of-a-vector-in-rust