Release Build Script
Going to automate releases to be published to GitHub when I make a new tag for a specific version. Sick.

This code is from GPT; "I want to have a GitHub Action for compiling my Rust CLI tool as an artifact in a GitHub release tag."

Going to modify this as necessary for any issues, or just to improve it in general. I haven't had much luck with completely understanding GitHub Actions so far, so I want to learn how to make good use of it.

https://doc.rust-lang.org/book/ch14-01-release-profiles.html
https://github.com/rust-build/rust-build.action (reference)
https://www.reddit.com/r/rust/comments/rrc9mv/rust_releasebuild_scripts/
https://github.com/SUPERCILEX/ftzz/blob/b68a7910e0da52ac7541f4671f2dd8d14c732bde/.github/workflows/cid.yml (reference)