Complete Build Targets!
Yay! Now all of the system build targets should be available through the GItHub Action. Hopefully this works.

It seems like the official name for ARM builds is aarch64, but most naming on sites and such show it as arm64, like when you download VSCode for example. So which do you think should be used? Like in terms of describing it to end-users? I don't want to make it more confusing than it already is.

https://dzfrias.dev/blog/deploy-rust-cross-platform-github-actions/
https://superuser.com/questions/1716835/on-which-systems-should-i-prefer-musl-or-gnu-binaries
https://github.com/cross-rs/cross?tab=readme-ov-file#supported-targets
https://stackoverflow.com/questions/31851611/differences-between-arm64-and-aarch64
https://www.reddit.com/r/asm/comments/10chhth/are_arm64_and_aarch64_the_same_or_thing/