aarch64 Linux Issues
This seems to be what GPT is recommending from, it appear to step from that running this on x86_64 Ubuntu, the cross-compilation for ARM Linux isn't installed yet, hopefully this addresses it properly. I really don't like doing a heck ton of commits back and forth like this, but I need to get this working, and learn what the issue is, and to just learn about all of this in general. So, it's okay that this is a mess haha. Ideally I only do this the once, then I can bring this workflow over to other Rust projects I make.

Listening to "Maggot Twat"

https://askubuntu.com/questions/1490387/how-do-i-install-the-gcc-13-aarch64-cross-compiler-on-ubuntu-22-04