Dir Reading Complete
Now I'm logging the contents of the files after having read them from the player folder.

I'm curious about the performance of file IO being sync here, compared to async like in JS/Node in general. Say if this were to be rewritten in JS, would it be faster because the file IO could be made into a single set of calls, rather than blocking?

https://stackoverflow.com/questions/39785597/how-do-i-get-a-slice-of-a-vect-in-rust (I think I had visited this before, forgot how to do it again)