Writing Support!
This worked successfully!

If you make any changes to `~local_player.nbt`, it will save them back into the DB when you run the script the next time again.

You can try this by changing the dimension ID for the player, then opening the world again. You should be in a different place than the Overworld, it is that by default in the world I have here, that's how I saved it last.