Complete Release
This essentially the alpha release, I just want to test more that the code works properly. I don't want to unintentionally break people's databases when running this tool. So the features themselves seem to work I think, it's just whether it's bug-free and no security issues as well.

I am thinking about abstracting this away into a nice GUI, probably driven by Tauri. I also could just make this more interactive too, to where the app would pause until the edits are done, and you would press resume once you wanted to write them back into the database. It could also ensure that the folder is less likely to be tampered with, since I would want to remove it after they have made their edits there.

Removed the ARM64 build for Linux, as it wasn't working properly with GitHub Actions. I will look into that again down the road, when I learn more about it, and in the event that I need to make pre-builds for that. I guess the RPi is a good example of needing to use something like this, I may want to figure it out specifically just for something like that.