Mojang Options
Found documentation and an example of how to use `rusty-leveldb` with a setup that allows for the compression types that Mojang uses for Bedrock! I think this means that `rusty-leveldb` might actually be fully compatible with what `leveldb-zlib` does in Node. This is great! Let alone having it all in Rust, but it might even allow for running this in WASM afterall, that's still hypothetical though, we'll see. More than anything else though, I love that the compression for this is fully plug-n-play, you don't have to change the code for `rusty-leveldb` itself to modify the compression it uses, you just pass in the list of ones you want to define, and what you want to write with.

https://github.com/dermesser/leveldb-rs/issues/45
https://github.com/dermesser/leveldb-rs/pull/46
https://github.com/dermesser/leveldb-rs/pull/46/commits/fcefa22c073e1c91428341d32f0f27d416d26f2b
https://github.com/dermesser/leveldb-rs/blob/master/examples/mcpe/src/main.rs ⭐️
https://github.com/dermesser/leveldb-rs/blob/master/examples/asyncdb-tokio/src/main.rs (yay! I was curious about if I could make the DB usage async)
https://github.com/Frommi/miniz_oxide
https://github.com/richgel999/miniz
https://docs.rust-embedded.org/book/intro/no-std.html (would this be helpful for WASM?)
https://doc.rust-lang.org/book/ch19-06-macros.html (nice learning material)